Remove `useDraftModePerspective`
Split from #3109
